+++ src/comp/MainPanel.js
@@ -1,6 +1,7 @@
 const vscode = require('vscode');
 const getNonce = require('./getNonce');
 const closeViewTracker = require('./closeCounter');
+const bencher = require('./bencher');
 //const sidebar = require('./SideBarProvider');
 
 
@@ -36,6 +37,7 @@ class MainPanel {
             // And restrict the webview to only loading content from our extension's `media` directory.
             localResourceRoots: [
                 vscode.Uri.joinPath(extensionUri, "media"),
+                vscode.Uri.joinPath(extensionUri, "out"),
                 vscode.Uri.joinPath(extensionUri, "out/compiled"),
             ],
         });
@@ -70,21 +72,20 @@ class MainPanel {
         this._panel.webview.html = this._getHtmlForWebview(webview);
         webview.onDidReceiveMessage(async (data) => {
             switch (data.type) {
+                case "benchmark": {
+                    if (!data.value) {
+                        return;
+                    }else{
+                        let results = bencher.runIt();
+                        console.log(results);
+                    }
+                    console.log("Sync HERE!");
+                    break;
+                }
                 case "firstSync": {
                     if (!data.value) {
                         return;
                     }
-                    //sidebar.SidebarProvider.
-                    /*
-                    var _a;
-					(_a = sidebar.SidebarProvider._view) === null || _a === void 0 ? void 0 : _a.webview.postMessage({
-						type: "new-function",
-						value: {
-							name: data.value.name,
-							id: data.value.id
-						},
-                    })
-                    */
                     console.log("Sync HERE!");
                     break;
                 }
@@ -114,6 +115,7 @@ class MainPanel {
         // Uri to load styles into webview
         //const styleResetUri = webview.asWebviewUri(styleResetPath);
         //const styleVSCodeUri = webview.asWebviewUri(stylesPathMainPath);
+        //const workerUri =  webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, "out", "working.js"));
         const styleMainUri = webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, "out", "compiled/WebPanel.css"));
         // Use a nonce to only allow specific scripts to be run
         const nonce = getNonce.getNonce();
@@ -129,13 +131,69 @@ class MainPanel {
           <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
 			integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
           <link href="${styleMainUri}" rel="stylesheet">
-          <script nonce="${nonce}">
+          <script id="API" nonce="${nonce}">
             const jsVscode = acquireVsCodeApi();
           </script>
       </head>
       <body>
+      <script type="text/js-worker">
+      // This script WON'T be parsed by JS engines because its MIME type is text/js-worker.
+      let myVar = 'Hello World! I AM A WORKER THAT IS OPPRESSED';
+      // Rest of your worker code goes here.
+      importScripts('https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js');
+      importScripts('https://cdnjs.cloudflare.com/ajax/libs/platform/1.3.6/platform.min.js');
+      importScripts('https://cdnjs.cloudflare.com/ajax/libs/benchmark/2.1.4/benchmark.min.js');
+      console.log(myVar);
+
+      const methods = {
+          fun1: function (x) {
+              console.log("Hello!!! Fun " + x);
+          },
+          fun2: function () {
+              console.log("Hello!! FUN2");
+          },
+      };
+
+      let suite = new Benchmark.Suite;
+      let results = [];
+      onmessage = function (oEvent) {
+          // add tests
+          suite.add('fun1', function () {
+                  let x = 10;
+                  methods.fun1(x);
+              })
+              .add('fun2', function () {
+                  methods.fun2();
+              })
+              // add listeners
+              .on('cycle', function (event) {
+                  console.log(String(event.target));
+              })
+              .on('complete', function () {
+                  console.log("Completed!");
+                  console.log('Fastest is ' + this.filter('fastest').map('name'));
+                  for (var i = 0; i < this.length; i++) {
+                      //console.log(this[i].hz + " ops/sec");
+                      //console.log(this[i].stats.rme + " Relative Margin of Error");
+                      let result = {
+                          ops: this[i].hz,
+                          rme: this[i].stats.rme
+                      }
+                      results.push(result);
+                  }
+              })
+              // run async
+              .run({
+                  'async': true
+              });
+          let returnVal = JSON.stringify(results);
+          postMessage(returnVal);
+      };
+  </script>
+
         <!-- Svelte Scripts below -->
-        <script nonce="${nonce}" src="${scriptUri}"></script>
+        <script nonce="${nonce}" id="svelte" src="${scriptUri}"></script>
+        
         <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
         integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous">
         </script>
@@ -154,6 +212,7 @@ MainPanel._panel = undefined;
 MainPanel.viewType = "main-panel";
 MainPanel.currentPanel = undefined;
 MainPanel.lastPanel = undefined;
+
 
 
 module.exports = {