+++ src/extension.js
@@ -1,22 +1,37 @@
 const vscode = require('vscode');
-//const handleBox = require('./comp/handleBox');
 const findMethod = require('./comp/findMethod');
 const sidebarProvider = require('./comp/SideBarProvider');
 const methodStorage = require('./comp/storeMethods');
-const mainDisplay = require('./comp/MainPanel');
-const closeViewTracker = require('./comp/closeCounter');
+//const mainDisplay = require('./comp/MainPanel');
+//const closeViewTracker = require('./comp/closeCounter');
+
+
+const open = require('open');
+const {instance} = require('./comp/objController/serverInstance');
+let browserOpened = false;
+
+
+
+//import {COMMAND, CONFIGURATION} from './localserver/config';
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 
 //To Mark when the WebView is closed
 
+//const webpage = new ResultsPanel.ResultsPanel();
+//const nodeServer = new node.LocalServer(webpage.getHtml());
+
 /**
  * @param {vscode.ExtensionContext} context
  */
 function activate(context) {
-
+	
 	const sbprov = new sidebarProvider.SidebarProvider(context.extensionUri);
-	context.subscriptions.push(vscode.window.registerWebviewViewProvider("benchMe_sidebar", sbprov))
+	context.subscriptions.push(vscode.window.registerWebviewViewProvider("benchMe_sidebar", sbprov));
+	instance.createServer();
+	
+
+	//currentServer = nodeServer;
 
 	console.log('Congratulations, your extension "benchme" is now active!');
 
@@ -75,11 +90,10 @@ function activate(context) {
 				setTimeout(function(){ 
 					 
 					findFunction(method, foundEditor).then((data) => {
-						//Function is found from here!
-
+						//Function is found from here! Start the server!
 						console.log(data);
 
-						//Sends to SideBar
+						//Sending to Sidebar
 						var _a;
 						(_a = sbprov._view) === null || _a === void 0 ? void 0 : _a.webview.postMessage({
 							type: "new-function",
@@ -94,7 +108,24 @@ function activate(context) {
 						methodStorage.pushToStore(methodInfo);
 						methodCounter++;
 
+						if(browserOpened){
+							instance.handleMsg({type: 'new-function',data: methodInfo});
+						}else{
+							open('http://localhost:52999').then(() =>{
+								//Send to new function to server
+								instance.handleMsg({type: 'new-function',data: methodInfo});
+								browserOpened = true;
+							}).catch((error)=>{
+								console.log(error);
+								console.log("Could not open Browser!");
+								vscode.window.showWarningMessage('Error! Could not open Browser!');
+							});
+						}
 
+						/**
+						*  This is all the WebView Compenent which is depcreated!
+						*  Unusable due to No WebWorker Support.
+						
 						let loadSaveData = closeViewTracker.getClose();
 
 						//Display Web View and send in the new data, whilst checking if there were any previous tracked functions
@@ -113,6 +144,16 @@ function activate(context) {
 								data: methodInfo
 							});
 						}
+						*/
+
+						//STARTS HERE ---------------
+						/*
+						setTimeout(function(){ 
+							const webpage = new ResultsPanel.ResultsPanel(context.extensionUri);
+							const nodeServer = new nodeserver.LocalServer(context.extensionUri,webpage.getHtml());
+							nodeServer.createServer();
+						},300);
+						*/
 
 						/*Switching Views!!
 						mainDisplay.MainPanel.currentPanel._panel.onDidChangeViewState(e => {
@@ -123,14 +164,14 @@ function activate(context) {
 
 					}).catch((errorMsg) => {
 						//Function not found!
-						vscode.window.showWarningMessage('Error! Function not found!');
+						vscode.window.showWarningMessage('Function not found!');
 						console.log("ERROR ON " + errorMsg);
 					})
 				}, 100);
 
 			}).catch((errorMsg) => {
 				//Active Editor Not found
-				vscode.window.showWarningMessage('Error! No Active Editor');
+				vscode.window.showWarningMessage('Warning! No Active Editor');
 				console.log("ERROR ON " + errorMsg);
 			})
 		}).catch((method) => {
@@ -147,8 +188,11 @@ exports.activate = activate;
 
 
 
-// this method is called when your extension is deactivated
+/**
+ * this method is called when your extension is deactivated
+ */
 function deactivate() {}
+