+++ webviews/pages/BrowserPage.js
@@ -1,4 +1,6 @@
-import App from "../components/WebPanel.svelte";
+import App from "../components/BrowserPage.svelte";
+
+//Ignore Error Below 
 const app = new App({
   // @ts-ignore
   // eslint-disable-next-line no-undef