+++ CHANGELOG.md
@@ -7,3 +7,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+## [1.0.0]
+- Benchmarking inside a Web Worker via default browser
+- Communication between default browser and extension via Web Sockets
+- Displaying results in the default browser
+- Sidebar functional TODO: Fix Retaining Information
+- TODO: Default Benchmarking Tests available and Live
+- TODO: Example Based Programming